requirements.txt
reportlab==3.5.68
img2pdf==0.4.1
pdfkit==0.6.1
Pillow==8.3.1
python-dotenv==0.18.0
argparse==1.4.0
pyTelegramBotAPI==3.8.2
aiogram==2.14.3
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
idna==3.2
multidict==5.1.0
pytz==2021.1
typing-extensions==3.10.0.0
yarl==1.6.3